add watchlist assets checking
