+++ src/components/SignalCard/card/SignalCard.module.scss
@@ -133,6 +133,10 @@
   }
 }
 
+.shareSingle {
+  padding: 0;
+}
+
 .popupButton {
   @include text('body-3');
