+++ src/ducks/Signals/signalFormManager/signalCrudForm/formParts/metricTypes/TriggerFormMetricTypes.js
@@ -1,18 +1,12 @@
-import React, { useState, useEffect } from 'react'
-import { compose } from 'recompose'
-import { connect } from 'react-redux'
+import React, { useState } from 'react'
 import Dialog from '@santiment-network/ui/Dialog'
 import {
   getNearestTypeByMetric,
   getSlugFromSignalTarget
 } from '../../../../utils/utils'
 import { METRICS_OPTIONS, TRENDING_WORDS } from '../../../../utils/constants'
-import { getCategoryGraph } from '../../../../../Studio/Sidebar/utils'
-import { Metric } from '../../../../../dataHub/metrics'
-import MetricsList from './MetricsList'
-import Search from './../../../../../Studio/Sidebar/Search'
 import MetricTypeRenderer from '../metricTypeRenderer/MetricTypeRenderer'
-import { withSignalMetrics } from '../../../../../Studio/withMetrics'
+import SupportedMetricsList from './SupportedMetricsList'
 import styles from '../../signal/TriggerForm.module.scss'
 import metricStyles from './TriggerFormMetricTypes.module.scss'
 
@@ -22,121 +16,16 @@ const checkPossibleTarget = ({ metaFormSettings, setFieldValue, target }) => {
   }
 }
 
-const makeSignalMetric = (key, label, category, node = 'line') => {
-  return {
-    key,
-    label,
-    category,
-    node
-  }
-}
-
-export const SIGNAL_SUPPORTED_METRICS = [
-  Metric.social_volume_total,
-  makeSignalMetric(
-    'social_volume_discord',
-    'Social volume (discord)',
-    'Social'
-  ),
-  makeSignalMetric(
-    'social_volume_professional_traders_chat',
-    'Social volume (pro traders chat)',
-    'Social'
-  ),
-  makeSignalMetric('social_volume_reddit', 'Social volume (reddit)', 'Social'),
-  makeSignalMetric(
-    'social_volume_telegram',
-    'Social volume (telegram)',
-    'Social'
-  ),
-  makeSignalMetric(
-    'social_volume_twitter',
-    'Social volume (twitter)',
-    'Social'
-  ),
-
-  Metric.volume_usd,
-  Metric.age_destroyed,
-  Metric.exchange_balance,
-  makeSignalMetric('price_btc', 'Price BTC', 'Financial'),
-  Metric.marketcap_usd,
-
-  makeSignalMetric(
-    'community_messages_count_total',
-    'Community messages count(total)',
-    'Social'
-  ),
-  makeSignalMetric(
-    'community_messages_count_telegram',
-    'Community messages count(telegram)',
-    'Social'
-  ),
-  // makeSignalMetric('community_messages_count_discord', 'Community messages count(discord)', 'Social'),
-
-  makeSignalMetric(
-    'social_dominance_total',
-    'Social dominance (total)',
-    'Social'
-  ),
-  makeSignalMetric(
-    'social_dominance_discord',
-    'Social dominance (discord)',
-    'Social'
-  ),
-  makeSignalMetric(
-    'social_dominance_professional_traders_chat',
-    'Social dominance (pro traders chat)',
-    'Social'
-  ),
-  makeSignalMetric(
-    'social_dominance_reddit',
-    'Social dominance (reddit)',
-    'Social'
-  ),
-  makeSignalMetric(
-    'social_dominance_telegram',
-    'Social dominance (telegram)',
-    'Social'
-  ),
-  makeSignalMetric(
-    'social_dominance_twitter',
-    'Social dominance (twitter)',
-    'Social'
-  ),
-
-  Metric.transaction_volume,
-  makeSignalMetric('exchange_inflow', 'Exchange Inflow', 'On-chain', 'bar'),
-  makeSignalMetric('exchange_outflow', 'Exchange Outflow', 'On-chain', 'bar'),
-  Metric.dev_activity,
-  makeSignalMetric('github_activity', 'Github Activity', 'Development')
-]
-
-const getByAvailable = availableMetrics =>
-  SIGNAL_SUPPORTED_METRICS.filter(({ key }) => {
-    return availableMetrics.indexOf(key) !== -1
-  })
-
 const TriggerFormMetricTypes = ({
   metric,
   target,
   setFieldValue,
   metaFormSettings,
-  availableMetrics
+  trigger
 }) => {
   const defaultMetric = metaFormSettings.metric
 
   const [open, setOpen] = useState(false)
-
-  const [categories, setCategories] = useState({})
-
-  useEffect(
-    () => {
-      const metrics = getByAvailable(availableMetrics)
-      const newCategories = getCategoryGraph(metrics, [])
-      setCategories(newCategories)
-    },
-    [target, availableMetrics]
-  )
 
   const onSelectMetric = newMetric => {
     metric &&
@@ -163,7 +52,7 @@ const TriggerFormMetricTypes = ({
     setOpen(false)
   }
 
-  const categoriesKeys = Object.keys(categories)
+  const slug = getSlugFromSignalTarget(trigger)
 
   return (
     <div className={styles.row}>
@@ -195,33 +84,7 @@ const TriggerFormMetricTypes = ({
               ))}
             </div>
 
-            <div className={metricStyles.choose}>
-              <div className={metricStyles.chooseText}>
-                or choose from the group of metrics
-              </div>
-              <div className={metricStyles.divider} />
-            </div>
-
-            <Search
-              iconPosition='left'
-              inputProps={{
-                placeholder: 'Search for a Metric'
-              }}
-              toggleMetric={onSelectMetric}
-              className={metricStyles.search}
-              categories={categories}
-            />
-
-            <div className={metricStyles.metrics}>
-              {categoriesKeys.map(key => (
-                <MetricsList
-                  key={key}
-                  metrikKey={key}
-                  list={categories[key]}
-                  onSelect={onSelectMetric}
-                />
-              ))}
-            </div>
+            <SupportedMetricsList slug={slug} onSelectMetric={onSelectMetric} />
           </div>
         </Dialog.ScrollContent>
       </Dialog>
@@ -229,11 +92,4 @@ const TriggerFormMetricTypes = ({
   )
 }
 
-const mapStateToProps = (state, { target }) => ({
-  slug: getSlugFromSignalTarget(target)
-})
-
-export default compose(
-  connect(mapStateToProps),
-  withSignalMetrics
-)(TriggerFormMetricTypes)
+export default TriggerFormMetricTypes
