+++ src/ducks/Signals/signalFormManager/signalCrudForm/signal/TriggerForm.js
@@ -220,6 +220,7 @@ export const TriggerForm = ({
                   setFieldValue={setFieldValue}
                   metric={metric}
                   target={target}
+                  trigger={mappedTrigger}
                 />
               )}
