+++ src/ducks/Signals/signalFormManager/signalCrudForm/signal/TriggerForm.module.scss
@@ -278,7 +278,7 @@ a {
 }
 
 .previewWithDiviver {
-  padding-top: 22px;
+  padding-top: 6px;
 }
 
 .checkboxBlock {
